+++ ouroboros-network-protocols/src/Ouroboros/Network/Protocol/LocalTxMonitor/Codec.hs
@@ -14,12 +14,15 @@ module Ouroboros.Network.Protocol.LocalTxMonitor.Codec
   , codecLocalTxMonitorId
   ) where
 
+import Control.Monad
 import Control.Monad.Class.MonadST
 
 import Network.TypedProtocol.Codec.CBOR
 import Network.TypedProtocol.Core
 
 import Data.ByteString.Lazy (ByteString)
+import Data.Map.Strict (Map)
+import Data.Map.Strict qualified as Map
 
 import Codec.CBOR.Decoding qualified as CBOR
 import Codec.CBOR.Encoding qualified as CBOR
@@ -64,6 +67,8 @@ codecLocalTxMonitor encodeTxId decodeTxId
         CBOR.encodeListLen 2 <> CBOR.encodeWord 7 <> encodeTxId txid
       MsgGetSizes ->
         CBOR.encodeListLen 1 <> CBOR.encodeWord 9
+      MsgGetMeasures ->
+        CBOR.encodeListLen 1 <> CBOR.encodeWord 11
       MsgAcquired slot ->
         CBOR.encodeListLen 2 <> CBOR.encodeWord 2 <> encodeSlot slot
       MsgReplyNextTx Nothing ->
@@ -79,6 +84,11 @@ codecLocalTxMonitor encodeTxId decodeTxId
         <> CBOR.encodeWord32 (capacityInBytes sz)
         <> CBOR.encodeWord32 (sizeInBytes sz)
         <> CBOR.encodeWord32 (numberOfTxs sz)
+      MsgReplyGetMeasures measures ->
+           CBOR.encodeListLen 3
+        <> CBOR.encodeWord 12
+        <> CBOR.encodeWord32 (txCount measures)
+        <> encodeMeasureMap (measuresMap measures)
 
     decode ::
          forall s (st :: ptcl).
@@ -105,6 +115,8 @@ codecLocalTxMonitor encodeTxId decodeTxId
           return (SomeMessage (MsgHasTx txid))
         (SingAcquired, 1, 9) ->
           return (SomeMessage MsgGetSizes)
+        (SingAcquired, 1, 11) ->
+          return (SomeMessage MsgGetMeasures)
 
         (SingAcquiring, 2, 2) -> do
           slot <- decodeSlot
@@ -128,11 +140,51 @@ codecLocalTxMonitor encodeTxId decodeTxId
           let sizes = MempoolSizeAndCapacity { capacityInBytes, sizeInBytes, numberOfTxs }
           return (SomeMessage (MsgReplyGetSizes sizes))
 
+        (SingBusy SingGetMeasures, 3, 12) -> do
+          txCount <- CBOR.decodeWord32
+          measuresMap <- decodeMeasureMap
+          let measures = MempoolMeasures { txCount, measuresMap }
+          pure (SomeMessage (MsgReplyGetMeasures measures))
+
         (SingDone, _, _) -> notActiveState stok
 
         (_, _, _) ->
           fail (printf "codecLocalTxMonitor (%s, %s) unexpected key (%d, %d)"
                        (show (activeAgency :: ActiveAgency st)) (show stok) key len)
+
+encodeMeasureMap :: Map MeasureName (SizeAndCapacity Integer) -> CBOR.Encoding
+encodeMeasureMap m =
+  CBOR.encodeMapLen (fromIntegral (Map.size m)) <>
+  Map.foldMapWithKey f m
+  where
+    f mn sc =
+      encodeMeasureName mn <> encodeSizeAndCapacity sc
+
+decodeMeasureMap :: CBOR.Decoder s (Map MeasureName (SizeAndCapacity Integer))
+decodeMeasureMap = do
+  len <- CBOR.decodeMapLen
+  mapContents <- replicateM len $
+    (,) <$> decodeMeasureName <*> decodeSizeAndCapacity
+  pure $ Map.fromList mapContents
+
+encodeMeasureName :: MeasureName -> CBOR.Encoding
+encodeMeasureName (MeasureName t) = CBOR.encodeString t
+
+decodeMeasureName :: CBOR.Decoder s MeasureName
+decodeMeasureName = MeasureName <$> CBOR.decodeString
+
+encodeSizeAndCapacity :: SizeAndCapacity Integer -> CBOR.Encoding
+encodeSizeAndCapacity sc =
+  CBOR.encodeListLen 2 <>
+  CBOR.encodeInteger (size sc) <>
+  CBOR.encodeInteger (capacity sc)
+
+decodeSizeAndCapacity :: CBOR.Decoder s (SizeAndCapacity Integer)
+decodeSizeAndCapacity = do
+  _len <- CBOR.decodeListLen
+  size <- CBOR.decodeInteger
+  capacity <- CBOR.decodeInteger
+  pure SizeAndCapacity { size, capacity }
 
 -- | An identity 'Codec' for the 'LocalTxMonitor' protocol. It does not do
 -- any serialisation. It keeps the typed messages, wrapped in 'AnyMessage'.
@@ -164,15 +216,19 @@ codecLocalTxMonitorId =
               => Message ptcl st st' -> m (DecodeStep bytes failure m (SomeMessage st))
           res msg = return (DecodeDone (SomeMessage msg) Nothing)
       in return $ DecodePartial $ \bytes -> case (stok, bytes) of
-        (SingIdle,     Just (AnyMessage msg@MsgAcquire{}))      -> res msg
-        (SingIdle,     Just (AnyMessage msg@MsgDone{}))         -> res msg
-        (SingAcquired, Just (AnyMessage msg@MsgAwaitAcquire{})) -> res msg
-        (SingAcquired, Just (AnyMessage msg@MsgNextTx{}))       -> res msg
-        (SingAcquired, Just (AnyMessage msg@MsgHasTx{}))        -> res msg
-        (SingAcquired, Just (AnyMessage msg@MsgRelease{}))      -> res msg
-        (SingAcquiring,       Just (AnyMessage msg@MsgAcquired{}))    -> res msg
-        (SingBusy SingNextTx, Just (AnyMessage msg@MsgReplyNextTx{})) -> res msg
-        (SingBusy SingHasTx,  Just (AnyMessage msg@MsgReplyHasTx{}))  -> res msg
+        (SingIdle,                 Just (AnyMessage msg@MsgAcquire{}))          -> res msg
+        (SingIdle,                 Just (AnyMessage msg@MsgDone{}))             -> res msg
+        (SingAcquired,             Just (AnyMessage msg@MsgAwaitAcquire{}))     -> res msg
+        (SingAcquired,             Just (AnyMessage msg@MsgNextTx{}))           -> res msg
+        (SingAcquired,             Just (AnyMessage msg@MsgHasTx{}))            -> res msg
+        (SingAcquired,             Just (AnyMessage msg@MsgRelease{}))          -> res msg
+        (SingAcquired,             Just (AnyMessage msg@MsgGetSizes{}))         -> res msg
+        (SingAcquired,             Just (AnyMessage msg@MsgGetMeasures{}))      -> res msg
+        (SingAcquiring,            Just (AnyMessage msg@MsgAcquired{}))         -> res msg
+        (SingBusy SingNextTx,      Just (AnyMessage msg@MsgReplyNextTx{}))      -> res msg
+        (SingBusy SingHasTx,       Just (AnyMessage msg@MsgReplyHasTx{}))       -> res msg
+        (SingBusy SingGetSizes,    Just (AnyMessage msg@MsgReplyGetSizes{}))    -> res msg
+        (SingBusy SingGetMeasures, Just (AnyMessage msg@MsgReplyGetMeasures{})) -> res msg
 
         (SingDone, _) -> notActiveState stok